pyproject.toml
[tool.poetry]
name = "agent-rag"
version = "0.1.0"
description = ""
authors = ["Elon Salfati <elon@salfati.group>","Andre Mayers"]
readme = "README.md"
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.3.14"
langgraph = "^0.2.61"
langchain-core = "^0.3.29"
langchain-community = "^0.3.14"
langchain-anthropic = "^0.3.1"
langchain-postgres = "^0.0.12"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
openai = "^1.59.5"
langchain-openai = "^0.2.14"
anthropic = "^0.42.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
typer = "^0.15.1"
rich = "^13.9.4"
unstructured = "^0.16.12"
networkx = "^3.4.2"
loguru = "^0.7.3"
psycopg-binary = "^3.2.3"
xlrd = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
